AccountRole entities, DTOs, Repository and Repository tests

+++ H3WebshopProeveprojekt.Api/Database/Entities/Product.cs
@@ -18,8 +18,8 @@ namespace H3WebshopProeveprojekt.Api.Database.Entities
         [Column(TypeName = "tinyint")]
         public int DiscountPercentage { get; set; }
 
-        [Column(TypeName = "smallint")]
-        public int CategoryId { get; set; }
+        [Column(TypeName = "int")]
+        public int CategoryId { get; set; } 
 
 
         public Category Category { get; set; }
